pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.7",
    "requests>=2.32.3",
    "setuptools>=76.0.0",
    "telegram>=0.0.1",
    "werkzeug>=3.0.1",
    "yfinance>=0.2.54",
]
